refactor: move encoding/utils/codec to encoding/codec
I want to split encoding into a v1 and v2 package. codec should be shared between both so pulling it out to the top level.
