+++ api/proxy/.env.example
@@ -44,7 +44,7 @@ EIGENDA_PROXY_EIGENDA_V2_NETWORK=sepolia_testnet
 # you only actually need to read the amount of SRS data that corresponds to the size of the largest blob that will be
 # sent, decreasing this value is a crude sort of optimization.
 # Note that this value is NOT the maximum payload size, because of encoding overhead.
-# See https://github.com/Layr-Labs/eigenda/blob/e5f489aae3/encoding/utils/codec/README.md for more details.
+# See https://github.com/Layr-Labs/eigenda/blob/e5f489aae3/encoding/codec/README.md for more details.
 EIGENDA_PROXY_EIGENDA_V2_MAX_BLOB_LENGTH=1MiB
 
 # Address of the EigenDACertVerifierRouter contract, or of an immutable EigenDACertVerifier contract of version >= 3.
